+++ src/editor/core/plugins/emoji/provider.js
@@ -67,7 +67,7 @@ let chooser = undefined;
 class EmojiChooser {
     constructor(provider) {
         this.provider = provider;
-        this.categoryOrder = ['people', 'animals_and_nature', 'food_and_drink', 'activity', 'travel_and_places', 'objects', 'symbols', 'flags'];
+        this.categoryOrder = ['people', 'animals_and_nature', 'food_and_drink', 'activity', 'travel_and_places', 'objects', 'symbols', 'flags', 'search'];
         this.categories = {
             people: {$icon: util.getCharToDom('\uD83D\uDE00')},
             animals_and_nature: {$icon: util.getCharToDom('\uD83D\uDC3B')},
@@ -76,7 +76,8 @@ class EmojiChooser {
             travel_and_places: {$icon: util.getCharToDom('\u2708\uFE0F')},
             objects: {$icon: util.getCharToDom('\uD83D\uDDA5')},
             symbols: {$icon: util.getCharToDom('\u2764\uFE0F')},
-            flags: {$icon: util.getCharToDom(findUserFlag())}
+            flags: {$icon: util.getCharToDom(findUserFlag())},
+            search: {$icon: util.getCharToDom('\uD83D\uDD0D')}
         };
     }
 
@@ -93,15 +94,70 @@ class EmojiChooser {
             top: position.top + provider.$node.outerHeight() - 5,
             left: position.left,
         }).show();
+
+        this.$.find('.humhub-emoji-chooser-search').focus();
     }
 
     initDom() {
         let that = this;
-        this.$ = $('<div class="atwho-view humhub-richtext-provider humhub-emoji-chooser">').hide().appendTo($('body')).on('hidden', () => {
-            if(that.provider) {
-                that.provider.reset();
+        this.$ = $('<div class="atwho-view humhub-richtext-provider humhub-emoji-chooser"><div><input type="text" class="form-control humhub-emoji-chooser-search"></div></div>')
+            .hide().appendTo($('body'))
+            .on('hidden', () => {
+                if(that.provider) {
+                    that.provider.reset();
+                }
+            });
+
+        this.$.find('.humhub-emoji-chooser-search').on('keydown', function(e) {
+            switch (e.which) {
+                case 9:
+                    e.preventDefault();
+                    that.nextCategory();
+                    break;
+                case 13:
+                    e.preventDefault();
+                    that.provider.select();
+                    break;
+                case 37:
+                    that.prev();
+                    break;
+                case 38:
+                    that.up();
+                    break;
+                case 39:
+                    that.next();
+                    break;
+                case 40:
+                    that.down();
+                    break;
             }
+        }).on('keyup', function(e) {
+            if (e.which !== 8 && !/[a-z0-9\d]/i.test(String.fromCharCode(e.which))) {
+                return;
+            }
+
+            let val = $(this).val();
+            if(!val.length && that.lastActiveCategory) {
+                that.openCategory(that.lastActiveCategory);
+                return;
+            }
+
+            if(val.length < 2) {
+                // TODO add this in next major version without clearsearch
+                /**$('[data-emoji-category="search"]').find('ul')
+                    .empty()
+                    .append('<li>Test</li>'); **/
+                return;
+            }
+
+            let currentlyActive = that.getActiveCategoryMenuItem().attr('data-emoji-nav-item');
+            if(currentlyActive !== 'search') {
+                that.lastActiveCategory = currentlyActive;
+            }
+
+            that.updateSearch(val);
         });
+
         this.initNav();
     }
 
@@ -121,12 +177,44 @@ class EmojiChooser {
 
             $nav.append($item);
         });
+
+        $nav.find('[data-emoji-nav-item="search"]').hide();
+    }
+
+    clearSearch() {
+        this.$.find('[data-emoji-nav-item="search"]').hide();
+        this.$.find('.humhub-emoji-chooser-search').val('');
+    }
+
+    updateSearch(searchStr) {
+        this.$.find('[data-emoji-nav-item="search"]').show();
+        let result = [];
+        this.categoryOrder.forEach((categoryName, index) => {
+            $.each(util.getByCategory(categoryName), (index, emoji) => {
+                if(emoji && emoji.keywords) {
+                    $.each(emoji.keywords, (index, keyword) => {
+                        if(keyword.includes(searchStr)) {
+                            result.push(emoji);
+                            return false;
+                        }
+                    });
+                }
+            });
+        });
+
+        this.openCategory('search');
+        this.setCategoryItems('search', result);
     }
 
     openCategory(categoryName) {
         let categoryDef = this.categories[categoryName];
-        if(!categoryDef.$) {
+
+        if(!this.$.find('[data-emoji-category="'+categoryName+'"]').length) {
             this.initCategory(categoryName)
+        }
+
+        if(categoryName !== 'search') {
+            this.clearSearch();
         }
 
         this.$.find('[data-emoji-nav-item]').removeClass('cur');
@@ -143,9 +231,23 @@ class EmojiChooser {
             that.provider.select();
         }).prependTo(this.$);
 
-        let $list = $('<ul class="atwo-view-ul">').appendTo($category);
-        let $li = undefined;
-        util.getByCategory(categoryName).forEach((emojiDef) => {
+        $('<ul class="atwo-view-ul humhub-emoji-chooser-item-list">').appendTo($category);
+        this.categories[categoryName].$ = $category;
+        this.setCategoryItems(categoryName);
+    }
+
+    setCategoryItems(categoryName, items) {
+        if(!items && categoryName !== 'search') {
+            items = util.getByCategory(categoryName);
+        }
+
+        if(!items) {
+            items = [];
+        }
+
+        let $list = this.categories[categoryName].$.find('.humhub-emoji-chooser-item-list').empty();
+
+        items.forEach((emojiDef) => {
             let $li = $('<li class="atwho-emoji-entry">').append(util.getCharToDom(emojiDef.char, emojiDef.name));
 
             if(categoryName === 'flags' && emojiDef.char === findUserFlag()) {
@@ -156,8 +258,6 @@ class EmojiChooser {
         });
 
         $list.children().first().addClass('cur');
-
-        this.categories[categoryName].$ = $category;
     }
 
     reset() {
@@ -185,6 +285,19 @@ class EmojiChooser {
 
     getActiveCategoryTab() {
         return this.$.find('[data-emoji-category]:visible');
+    }
+
+    getActiveCategoryMenuItem() {
+        return this.$.find('[data-emoji-nav-item].cur');
+    }
+
+    nextCategory() {
+        let $next = this.getActiveCategoryMenuItem().next('[data-emoji-nav-item]:not([data-emoji-nav-item="search"])');
+        if(!$next.length) {
+            $next = this.$.find('[data-emoji-nav-item]:first');
+        }
+
+        this.openCategory($next.attr('data-emoji-nav-item'));
     }
 
     prev() {